extract controllers for check-guess views
This is a way more sane way to design in IB, plus less juggling with hiding/showing views.

Also got the new fonts fitting :)
